% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{pasilla}
\alias{pasilla}
\title{Pasilla knock down experiment}
\format{An object of class \code{list} of length 3.}
\source{
\url{https://www.bioconductor.org/packages/release/data/experiment/html/pasilla.html}
}
\usage{
pasilla
}
\description{
Pasilla knock down experiment
}
\keyword{datasets}
