% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{rat}
\alias{rat}
\title{RNA-Seq: Rat neuron experiment}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed?term=20452967}
}
\usage{
rat
}
\description{
RNA-Seq: Rat neuron experiment
}
\keyword{datasets}
