% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{tissues}
\alias{tissues}
\title{RNA-Seq: Human Tissues Experiment}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.ncbi.nlm.nih.gov/pubmed/22496456}
}
\usage{
tissues
}
\description{
RNA-Seq: Human Tissues Experiment
}
\keyword{datasets}
