% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{RNA-Seq: Fission yeast mutant time series experiment}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.bioconductor.org/packages/release/data/experiment/html/fission.html}
}
\usage{
yeast
}
\description{
RNA-Seq: Fission yeast mutant time series experiment
}
\keyword{datasets}
