% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{zebrafish}
\alias{zebrafish}
\title{RNA-Seq: Zebrafish gallein experiment}
\format{An object of class \code{list} of length 3.}
\source{
\url{http://www.bioconductor.org/packages/release/data/experiment/html/zebrafishRNASeq.html}
}
\usage{
zebrafish
}
\description{
RNA-Seq: Zebrafish gallein experiment
}
\keyword{datasets}
