% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{lmExamples}
\alias{lmExamples}
\title{Example gene expression for common designs}
\format{An object of class \code{list} of length 6.}
\usage{
lmExamples
}
\description{
Example gene expression for common designs
}
\keyword{datasets}
